meta: rename decorator-utils
